Multiple pattern length reporting
Program now reports on a range of pattern lengths as specified by user.
User can specify the total number of results, as well as min and max
pattern length to search for.
Also have now excluded reporting on patterns which are all zero, as
these tended to generally take up the top positions (perhaps as should
be expected), and aren't of particular interest. Future versions should
have an option to choose to exclude or include zero patterns.