Fixed error with KMP algorithm
KMP algorithm was stepping through blocks of contiguous characters by
one step too far causing a miscount of patterns. KMP also seemed to have
been designed to step ahead the length of the search pattern if the
pattern was all one character. Changed to count all occurrences of
pattern in line with others.